🏗 Co-routines yurs trully

+++ from-paris-to-berlin-data/src/main/kotlin/org/jesperancinha/fptb/circuit/breaker/domain/RoadBlockTime.kt
@@ -4,6 +4,6 @@ package org.jesperancinha.fptb.circuit.breaker.domain
  * Created by jofisaes on 14/10/2021
  */
 data class RoadBlockTime(
-    val minute: Long,
+    val minute: Int,
     val blockageType: BlockageType
 )
